Added class enrollment functionality, added 2 users (1 student 1 admin), 5 courses, and 3 courses enrolled in by the student
Server.js - app.get('/enrollCourses', (req, res) creates a list of the courses that a student is not enrolled in and makes a checkbox form on the view_logged for them to select what class to enroll in.
app.post('/enroll_action'... - takes the form submission, determines if its a list of submissions or just one, then calls findPromise on each course to enroll the student in them.
source.js - enrollCourses() fetches the enrollCourse get to create the form of possible courses to enroll in
view_logged - added the shell for the form